+++ packages/headless/src/features/facets/facet-set/interfaces/options.ts
@@ -1,16 +1,15 @@
 import {FacetRequest} from './request';
 
 type FacetRequiredParameters = Pick<FacetRequest, 'facetId' | 'field'>;
-type FacetOptionalParameters = Partial<
-  Pick<
-    FacetRequest,
-    | 'delimitingCharacter'
-    | 'filterFacetCount'
-    | 'injectionDepth'
-    | 'numberOfValues'
-    | 'sortCriteria'
-  >
+
+export type FacetOptionalParameters = Pick<
+  FacetRequest,
+  | 'delimitingCharacter'
+  | 'filterFacetCount'
+  | 'injectionDepth'
+  | 'numberOfValues'
+  | 'sortCriteria'
 >;
 
 export type FacetRegistrationOptions = FacetRequiredParameters &
-  FacetOptionalParameters;
+  Partial<FacetOptionalParameters>;